Return a DENIED permissionsType instead of raising RemoteError Denied.
This is in response to the asTest rpc.
When the PV is denied, the current response "Error: Denied"
is misleading as Denied is a valid response to the query.
Also, returning a permissionsType structure provides confirmation
to the user of which user, host, roles, etc were used in the query.